refactor(logging): move sentry config
